将user表和userInfo表合并

+++ ppeng-module-user/src/main/java/fun/zhub/ppeng/dto/update/UserInfoUpdateDTO.java
@@ -53,7 +53,8 @@ public class UserInfoUpdateDTO implements Serializable {
     /**
      * 性别
      */
-    private Byte gender;
+
+    private Integer gender;
 
     /**
      * 生日
